python-service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pandas==2.3.2
numpy==2.3.2
